+++ src/main/kotlin/backend/itracker/tracker/service/handler/MacbookHandler.kt
@@ -1,4 +1,4 @@
-package backend.itracker.tracker.service
+package backend.itracker.tracker.service.handler
 
 import backend.itracker.crawl.common.ProductCategory
 import backend.itracker.crawl.macbook.service.MacbookService
